added messages to forms and cleaned up the code a bit

+++ PMT/dashboard/views.py
@@ -1,5 +1,6 @@
 from django.shortcuts import render, redirect
 from django.contrib.auth.decorators import login_required
+from django.contrib import messages
 
 from . import discovery_algorithms
 from .forms import EventLogForm, ProcessModelForm, DiscoverProcessModelForm
@@ -18,6 +19,10 @@ def index(request):
 def process_discovery(request):
     """Function for process discovery"""
     process_models = ProcessModel.objects.all()
+    template = "dashboard/process_discovery.html"
+    redirect_url = "/process_discovery"
+    selected_process_model = None
+    context = {}
     if request.method == "POST":
         discover_process_form = DiscoverProcessModelForm(request.POST)  # Get form model instance
         if "submitDiscover" in request.POST:  # Discovery algorithms to produce process model file and png
@@ -28,77 +33,105 @@ def process_discovery(request):
                 logfile = str(new_process_model.process_model_log_name)
                 process_model_name = new_process_model.process_model_name
                 discovery_algorithm = new_process_model.process_model_algorithm
-
                 # Apply algorithm based on user selection and update model fields
                 if discovery_algorithm == "alpha_miner":
-                    discovery_algorithms.discovery_alpha_miner(logfile, process_model_name)
-                    new_process_model.process_model_file = "process_models/" + process_model_name + ".pnml"
-                    new_process_model.process_model_image = "exported_pngs/" + process_model_name + ".png"
+                    if discovery_algorithms.discovery_alpha_miner(logfile, process_model_name):
+                        messages.success(request, "Process Model discovered successfully!")
+                        new_process_model.process_model_file = "process_models/" + process_model_name + ".pnml"
+                        new_process_model.process_model_image = "exported_pngs/" + process_model_name + ".png"
+                    else:
+                        messages.error(request, "Something went wrong! Please try again.")
+                        redirect(redirect_url)
+
                 elif discovery_algorithm == "inductive_miner":
                     pass
                 elif discovery_algorithm == "heuristic_miner":
                     pass
-
                 # Save the process model
                 new_process_model.save()
                 discover_process_form.save_m2m()
-
-                return redirect("/process_discovery")
+                return redirect(redirect_url)
         if "submitRender" in request.POST:  # Get selected process model and return it
             process_model_id = request.POST["process_model_id"]
             selected_process_model = ProcessModel.objects.get(pk=process_model_id)
-            return render(request, "dashboard/process_discovery.html", {"discover_process_form": discover_process_form,
-                                                                        "process_models": process_models,
-                                                                        "selected_process_model": selected_process_model})
     else:
         discover_process_form = DiscoverProcessModelForm()
-
-    return render(request, "dashboard/process_discovery.html", {"discover_process_form": discover_process_form,
-                                                                "process_models": process_models})
+    context["discover_process_form"] = discover_process_form
+    context["process_models"] = process_models
+    if selected_process_model is not None:
+        context["selected_process_model"] = selected_process_model
+    return render(request, template, context)
 
 
 @login_required(login_url="/accounts/login")
 def view_process_models(request):
     """Function that renders view processes"""
+    redirect_url = "/view_process_models"
+    template = "dashboard/view_process_models.html"
+    context = {}
     process_models = ProcessModel.objects.all()
     if request.method == "POST":
         form = ProcessModelForm(request.POST, request.FILES)
         if form.is_valid():
-            form.save()
-            return redirect("/view_process_models")
+            if form.save():
+                messages.success(request, "Process Model added successfully!")
+            else:
+                messages.error(request, "Something went wrong! Please try again!")
+            return redirect(redirect_url)
     else:
         form = ProcessModelForm()
-    return render(request, "dashboard/view_process_models.html", {"form": form, "process_models": process_models})
+    context["form"] = form
+    context["process_models"] = process_models
+    return render(request, template, context)
 
 
 @login_required(login_url="/accounts/login")
 def delete_process_model(request, pk):
     """Function to handle deletion of event logs"""
+    redirect_url = "/view_process_models"
     if request.method == "POST":
         process_model = ProcessModel.objects.get(pk=pk)
-        process_model.delete()
-    return redirect("/view_process_models")
+        if process_model is None:
+            messages.error(request, "Process Model not found!")
+        else:
+            process_model.delete()
+            messages.success(request, "Process Model deleted successfully!")
+    return redirect(redirect_url)
 
 
 @login_required(login_url="/accounts/login")
 def event_logs(request):
     """Function for event logs management"""
     logs = EventLog.objects.all()
+    template = "dashboard/event_logs.html"
+    context = {}
     if request.method == "POST":
         form = EventLogForm(request.POST, request.FILES)
         if form.is_valid():
-            form.save()
+            if form.save():
+                messages.success(request, "Event log added successfully!")
+            else:
+                messages.error(request, "Something went wrong! Please try again.")
+
             return redirect("/event_logs")
     else:
         form = EventLogForm()
 
-    return render(request, "dashboard/event_logs.html", {"form": form, "event_logs": logs})
+    # Add items to context
+    context["form"] = form
+    context["event_logs"] = logs
+    return render(request, template, context)
 
 
 @login_required(login_url="/accounts/login")
 def delete_event_log(request, pk):
     """Function to handle deletion of event logs"""
+    redirect_page = "/event_logs"
     if request.method == "POST":
         event_log = EventLog.objects.get(pk=pk)
-        event_log.delete()
-    return redirect("/event_logs")
+        if event_log is None:
+            messages.error(request, "Event Log not found!")
+        else:
+            event_log.delete()
+            messages.success(request, "Event Log deleted successfully!")
+    return redirect(redirect_page)
